completed breakout of components
